bs_usersテーブル構造修正

+++ src/database/migrations/2019_08_03_060155_create_bs_users_table.php
@@ -15,7 +15,7 @@ class CreateBsUsersTable extends Migration
     {
         Schema::create('bs_users', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('fb_id');
+            $table->bigInteger('fb_id');
             $table->timestamp('craeted_at');
             $table->softDeletes();
         });
